Move sklearn wrappers to a new file
